[k1c] Initial load/store packing pre-register allocation pass
Summary:
Packs LD and SD to LQ/LO, SQ/SO respectively. Load implementation checks the
entire block and tries to minimise to instructions as much as it can, for now it
checks if there are enough virtual registers (virtual registers are considered
physical which is not optimal in this case). Store for now can only be packed if
a load packing was done on the same registers, because of how llvm handles the
definition of subregisters.

Test Plan: LLVM CI

Reviewers: O81 llvm-project!

Differential Revision: https://phab.kalray.eu/D1657