[kvx] Fix bitcode raw attribute 'nofree'
Summary:
We can ignore the nofree atribute, as it
is used mostly for optimization issues,
and return None instead.
A better option is use one of the
attributes not used at all for Kalray
such as swift ones.

Using llvm_unreachable is UNPREDICTABLE
in release builds, it might not trigger
an error and generate incorrect code.
For some reason, the tests:
 - Bitcode/highLevelStructure.3.2.ll
 - Bitcode/miscInstructions.3.2.ll
 - Bitcode/standardCIntrinsic.3.2.ll
 - Bitcode/variableArgumentIntrinsic.3.2.ll
did not fail before, but with changes in
non-related code, such as adding some
extended inline asm support, makes these
tests to fail.

Test Plan: these 4 tests should not fail

Reviewers: O81 llvm-coolidge, mpoulhies

Reviewed By: O81 llvm-coolidge, mpoulhies

Differential Revision: https://phab.kalray.eu/D9029