[kvx] Refactor GPR registers definitions
Summary:
This patch changes how we define GPR registers, by always declaring
it is a pair of smaller types.
The gains of using named sub-indexes are such that we can easely
instanciate Paired or Single registers from a Quad, or compose
the larger one with any of the smaller ones. Doing so improves
codegen for vector elements moving.

Test Plan: added tests, CI.

Reviewers: lthevenoux, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D9154