[kvx][loopdo] Enable the IR passes
Summary:
This implements all the missing pieces to
activate the IR HardwareLoops.
First, the KVXHardwareLoopsPrepare pass,
that simplifies the lowering of the loop
end intrinsic, as it produces an i1 value,
which is illegal.

Second, KVXTargetMachine adds the call
for both the IR passes,
 - HardwareLoops
 - KVXHardwareLoopsPrepare

as well backend pass:
 - KVXHardwareLoops

Also we add some hidden flags to limit
the number of hwloops generated in a
module, useful to debug when required
to identify a faulty hwloop.

ref: T12703

Test Plan: added/existing tests

Reviewers: O81 llvm-coolidge, lthevenoux

Reviewed By: O81 llvm-coolidge, lthevenoux

Subscribers: lthevenoux

Maniphest Tasks: T12703

Differential Revision: https://phab.kalray.eu/D10447