[kvx] Add patterns for COMPUWD COMPWD
Summary:
This adds patterns of the word x double word integer comparison, in reg.
For immediates, it improves the usage of compw. When we have a [sz]ext
of an value, compared to an i64 immediate that fits in 32 bits, it
truncates the immediate, respecting the sing extension type.

Depends on D10567

Test Plan: added tests

Reviewers: O81 llvm-coolidge, lthevenoux, mpoulhies

Reviewed By: O81 llvm-coolidge, lthevenoux

Differential Revision: https://phab.kalray.eu/D10571